app/smc-image-upload/requirements.txt
aioboto3
aiohttp
pillow
requests_toolbelt